CC - New/requirements.txt
absl-py==2.1.0
alembic==1.13.1
astunparse==1.6.3
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.3
Flask-SQLAlchemy==3.0.3
flatbuffers==24.3.25
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
h5py==3.11.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.3.3
libclang==18.1.1
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.8
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pillow==10.3.0
protobuf==4.25.3
Pygments==2.18.0
PyJWT==2.8.0
requests==2.32.3
rich==13.7.1
setuptools==70.0.0
six==1.16.0
SQLAlchemy==2.0.31
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
timedelta==2020.12.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
